requirements.txt
# Core ML libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# MLflow for experiment tracking
mlflow>=2.14.0

# Configuration
PyYAML>=6.0.0

# Notebook dependencies (optional)
seaborn>=0.13.0
matplotlib>=3.8.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0